backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
groq
python-dotenv
pydantic
supabase